pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-tenancy"
version = "0.1.0"
description = "Enterprise-grade multi-tenancy for FastAPI — schema, database, RLS, and hybrid isolation with full async support"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "fastapi-tenancy contributors" }]
keywords = ["fastapi", "multitenancy", "saas", "postgresql", "sqlalchemy", "async"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "fastapi>=0.111.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "anyio>=4.0.0",
]

[project.optional-dependencies]
# Async database drivers (install the one matching your database)
postgres  = ["asyncpg>=0.29.0"]
sqlite    = ["aiosqlite>=0.19.0"]
mysql     = ["aiomysql>=0.2.0"]
mssql     = ["aioodbc>=0.5.0"]

# Optional services
redis      = ["redis[hiredis]>=5.0.0"]
jwt        = ["python-jose[cryptography]>=3.3.0"]
migrations = ["alembic>=1.13.0"]

# Convenience bundles
full = [
    "fastapi-tenancy[postgres,redis,jwt,migrations]",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-section-index>=0.3.0",
]
dev = [
    # all optional service extras — must be installed for the full test suite
    "fastapi-tenancy[postgres,sqlite,redis,jwt,migrations]",
    # tox matrix runner
    "tox>=4.36.0",
    # test stack
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "anyio[trio]>=4.0.0",
    "httpx>=0.27.0",
    # lint / type check
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-python-jose>=3.5",
    "sqlalchemy[mypy]>=2.0.30",
]

[project.urls]
Homepage      = "https://github.com/KapilDagur/fastapi-tenancy"
Documentation = "https://fastapi-tenancy.readthedocs.io"
Repository    = "https://github.com/KapilDagur/fastapi-tenancy"
Changelog     = "https://fastapi-tenancy.readthedocs.io/changelog/"
"Bug Tracker" = "https://github.com/KapilDagur/fastapi-tenancy/issues"

# Build
[tool.hatch.build.targets.wheel]
packages = ["src/fastapi_tenancy"]

[tool.hatch.build.targets.wheel.force-include]
"src/fastapi_tenancy/py.typed" = "fastapi_tenancy/py.typed"

# Version is the single source of truth in pyproject.toml.
# __init__.py reads it via importlib.metadata.version("fastapi-tenancy").
# Do NOT add a [tool.hatch.version] section — that would create a second source.

# pytest
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--tb=short",
    "--strict-markers",
    "-q",
]
markers = [
    "unit: Fast unit tests with no I/O",
    "integration: Integration tests (SQLite / mocks, no external services)",
    "e2e: End-to-end tests (may start a live database)",
    "slow: Tests that take > 1 s",
]

# coverage
[tool.coverage.run]
source = ["src/fastapi_tenancy"]
branch = true
omit   = ["*/tests/*", "*/__main__.py"]

[tool.coverage.report]
fail_under    = 95
show_missing  = true
skip_covered  = false
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
    "def __repr__",
]

# mypy
[tool.mypy]
python_version      = "3.11"
strict              = true
warn_return_any     = true
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = false
plugins             = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
exclude             = ["tests/", "docs/", "migrations/"]

[[tool.mypy.overrides]]
module = ["fastapi_tenancy.*"]
warn_return_any = false

# ruff─
[tool.ruff]
line-length    = 100
target-version = "py311"
src            = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",        # pyflakes
    "I",        # isort
    "B",        # flake8-bugbear
    "C4",       # comprehensions
    "UP",       # pyupgrade
    "SIM",      # simplify
    "TCH",      # type-checking imports
    "RUF",      # ruff-specific
    "S",        # bandit security
    "ANN",      # annotations
    "PT",       # pytest style
]
ignore = [
    "ANN101",  # self annotation
    "ANN102",  # cls annotation
    "ANN401",  # Any in annotations
    "S101",    # assert in tests
    "S106",    # hardcoded password in fixtures
    "B008",    # FastAPI Depends() calls in defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S", "PT011"]

[tool.ruff.format]
quote-style  = "double"
indent-style = "space"
